agenda/requirements.txt
backports.ssl==0.0.9
-e git+git@github.com:bricewge/42-agenda.git@9e13622a94b78e1ec7581e2f4267933330717eec#egg=caldav&subdirectory=venv/src/caldav
cffi==1.9.1
coverage==4.3.1
cryptography==1.7.1
enum34==1.1.6
funcsigs==1.0.2
idna==2.2
IMAPClient==1.0.2
ipaddress==1.0.17
lxml==3.7.1
mock==1.3.0
nose==1.3.7
pbr==1.10.0
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.12.4
six==1.10.0
vobject==0.9.3
